Add demos for boat endpoints

+++ server.js
@@ -1,5 +1,4 @@
 const express = require("express");
-const serveStatic = require("serve-static");
 const { expressjwt: jwt } = require("express-jwt");
 const jwksRsa = require("jwks-rsa");
 const { auth, requiresAuth } = require("express-openid-connect");
@@ -24,7 +23,7 @@ const {
 const config = {
   authRequired: false,
   auth0Logout: true,
-  baseURL: `${process.env.APP_URL}`,
+  baseURL: process.env.APP_URL,
   clientID: process.env.CLIENT_ID,
   issuerBaseURL: `https://${process.env.AUTH0_DOMAIN}`,
   secret: process.env.CLIENT_SECRET,
@@ -32,12 +31,6 @@ const config = {
 
 const app = express();
 
-// app.use("ship-solid.svg", express.static("public/ship-solid.svg"));
-// app.use(express.static(path.join(__dirname, "public")));
-app.use(
-  // serveStatic(path.join(__dirname, "public"), { index: ["welcome.html"] })
-  serveStatic(path.join(__dirname, "public"))
-);
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 app.use(auth(config));
@@ -64,34 +57,27 @@ const checkJwt = jwt({
 
 app.get("/", (req, res) => {
   if (req.oidc.isAuthenticated()) {
-    if (!req.accepts("application/json")) {
-      res.status(406).send({ Error: "server can only send application/json" });
-      return;
-    } else {
-      const userData = {
-        sub: req.oidc.user.sub,
-        name: req.oidc.user.name,
-        updated_at: req.oidc.user.updated_at,
-      };
-      userLogin(userData)
-        .then((user_id) => {
-          /* const userObj = { token: req.oidc.idToken, user_id: user_id };
-          res.json(userObj); */
-          const options = { root: path.join(__dirname) };
-          res.sendFile("public/main.html", options);
-        })
-        .catch((err) => {
-          console.error(err);
-          res.status(500).send(err);
+    const userData = {
+      sub: req.oidc.user.sub,
+      name: req.oidc.user.name,
+      updated_at: req.oidc.user.updated_at,
+    };
+    userLogin(userData)
+      .then(() => {
+        res.cookie("token", req.oidc.idToken, {
+          sameSite: "lax",
         });
-    }
+        const options = { root: path.join(__dirname) };
+        res.sendFile("public/main.html", options);
+      })
+      .catch((err) => {
+        console.error(err);
+        res.status(500).send(err);
+      });
   } else {
     const options = { root: path.join(__dirname) };
     res.sendFile("public/index.html", options);
   }
-  // const options = { root: path.join(__dirname) };
-  // res.set("Content-Type", "text/css");
-  // res.sendFile("public/main.html", options);
 });
 
 app.get("/profile", requiresAuth(), (req, res) => {
@@ -104,16 +90,6 @@ app.get("/profile", requiresAuth(), (req, res) => {
 
 app.get("/callback", (req, res) => {
   console.log("callback");
-});
-
-app.get("/main.html", (req, res) => {
-  if (req.oidc.isAuthenticated()) {
-    const options = { root: path.join(__dirname) };
-    res.sendFile("public/main.html", options);
-  } else {
-    const options = { root: path.join(__dirname) };
-    res.sendFile("public/index.html", options);
-  }
 });
 
 /* ------------- /users ------------- */
@@ -160,7 +136,6 @@ app.get("/boats", checkJwt, (req, res) => {
     res.status(401).send({ Error: "not authorized" });
   } else if (!req.accepts("application/json")) {
     res.status(406).send({ Error: "server can only send application/json" });
-    return;
   } else {
     let cursor = "";
     if (Object.keys(req.query).includes("cursor")) {
@@ -301,8 +276,10 @@ app.delete("/boats/:boat_id", checkJwt, (req, res) => {
 
 /* ------------- /cargo ------------- */
 
-app.get("/cargo", (req, res) => {
-  if (!req.accepts("application/json")) {
+app.get("/cargo", checkJwt, (req, res) => {
+  if (!req.auth) {
+    res.status(401).send({ Error: "not authorized" });
+  } else if (!req.accepts("application/json")) {
     res.status(406).send({ Error: "server can only send application/json" });
     return;
   } else {
@@ -321,8 +298,10 @@ app.get("/cargo", (req, res) => {
   }
 });
 
-app.post("/cargo", (req, res) => {
-  if (!req.accepts("application/json")) {
+app.post("/cargo", checkJwt, (req, res) => {
+  if (!req.auth) {
+    res.status(401).send({ Error: "not authorized" });
+  } else if (!req.accepts("application/json")) {
     res.status(406).send({ Error: "server can only send application/json" });
   } else if (!req.body.item || !req.body.volume) {
     res.status(400).json({
@@ -358,8 +337,10 @@ app.delete("/cargo", (req, res) => {
   res.status(405).send();
 });
 
-app.get("/cargo/:cargo_id", (req, res) => {
-  if (!req.accepts("application/json")) {
+app.get("/cargo/:cargo_id", checkJwt, (req, res) => {
+  if (!req.auth) {
+    res.status(401).send({ Error: "not authorized" });
+  } else if (!req.accepts("application/json")) {
     res.status(406).send({ Error: "server can only send application/json" });
     return;
   } else {
@@ -374,8 +355,10 @@ app.get("/cargo/:cargo_id", (req, res) => {
   }
 });
 
-app.patch("/cargo/:cargo_id", (req, res) => {
-  if (!req.accepts("application/json")) {
+app.patch("/cargo/:cargo_id", checkJwt, (req, res) => {
+  if (!req.auth) {
+    res.status(401).send({ Error: "not authorized" });
+  } else if (!req.accepts("application/json")) {
     res.status(406).send({ Error: "server can only send application/json" });
   } else {
     const bodyKeys = Object.keys(req.body);
@@ -400,8 +383,10 @@ app.patch("/cargo/:cargo_id", (req, res) => {
   }
 });
 
-app.put("/cargo/:cargo_id", (req, res) => {
-  if (!req.accepts("application/json")) {
+app.put("/cargo/:cargo_id", checkJwt, (req, res) => {
+  if (!req.auth) {
+    res.status(401).send({ Error: "not authorized" });
+  } else if (!req.accepts("application/json")) {
     res.status(406).send({ Error: "server can only send application/json" });
   } else if (!req.body.item || !req.body.volume) {
     res.status(400).json({
@@ -421,40 +406,59 @@ app.put("/cargo/:cargo_id", (req, res) => {
   }
 });
 
-app.delete("/cargo/:cargo_id", (req, res) => {
-  deleteCargo(parseInt(req.params.cargo_id))
-    .then((result) => {
-      res.status(result[0]).json(result[1]);
-    })
-    .catch((err) => {
-      console.error(err);
-    });
+app.delete("/cargo/:cargo_id", checkJwt, (req, res) => {
+  if (!req.auth) {
+    res.status(401).send({ Error: "not authorized" });
+  } else {
+    deleteCargo(parseInt(req.params.cargo_id))
+      .then((result) => {
+        res.status(result[0]).json(result[1]);
+      })
+      .catch((err) => {
+        console.error(err);
+      });
+  }
 });
 
 /* ------------- /boats/:boat_id/cargo/:cargo_id ------------- */
 
-app.put("/boats/:boat_id/cargo/:cargo_id", (req, res) => {
-  assignCargotoBoat(parseInt(req.params.boat_id), parseInt(req.params.cargo_id))
-    .then((result) => {
-      res.status(result[0]).json(result[1]);
-    })
-    .catch((err) => {
-      console.error(err);
-    });
+app.put("/boats/:boat_id/cargo/:cargo_id", checkJwt, (req, res) => {
+  if (!req.auth) {
+    res.status(401).send({ Error: "not authorized" });
+  } else {
+    assignCargotoBoat(
+      parseInt(req.params.boat_id),
+      parseInt(req.params.cargo_id)
+    )
+      .then((result) => {
+        res.status(result[0]).json(result[1]);
+      })
+      .catch((err) => {
+        console.error(err);
+      });
+  }
 });
 
-app.delete("/boats/:boat_id/cargo/:cargo_id", (req, res) => {
-  deleteCargotoBoatMapping(
-    parseInt(req.params.boat_id),
-    parseInt(req.params.cargo_id)
-  )
-    .then((result) => {
-      res.status(result[0]).json(result[1]);
-    })
-    .catch((err) => {
-      console.error(err);
-    });
+app.delete("/boats/:boat_id/cargo/:cargo_id", checkJwt, (req, res) => {
+  if (!req.auth) {
+    res.status(401).send({ Error: "not authorized" });
+  } else {
+    deleteCargotoBoatMapping(
+      parseInt(req.params.boat_id),
+      parseInt(req.params.cargo_id)
+    )
+      .then((result) => {
+        res.status(result[0]).json(result[1]);
+      })
+      .catch((err) => {
+        console.error(err);
+      });
+  }
 });
+
+/* ------------ Static files ------------ */
+
+app.use(express.static(path.join(__dirname, "public")));
 
 /* ------------- Start app ------------- */
